refactor: 도메인 모델 클래스 패키지 이동
