+++ android/app/src/main/java/com/mulkkam/data/repository/DevicesRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.mulkkam.data.remote.model.error.toDomain
 import com.mulkkam.data.remote.model.error.toResponseError
 import com.mulkkam.data.remote.model.request.DeviceRequest
 import com.mulkkam.data.remote.service.DevicesService
-import com.mulkkam.domain.MulKkamResult
+import com.mulkkam.domain.model.MulKkamResult
 import com.mulkkam.domain.repository.DevicesRepository
 
 class DevicesRepositoryImpl(
