+++ android/app/src/main/java/com/mulkkam/data/repository/NicknameRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.mulkkam.data.repository
 import com.mulkkam.data.remote.model.error.toDomain
 import com.mulkkam.data.remote.model.error.toResponseError
 import com.mulkkam.data.remote.service.NicknameService
-import com.mulkkam.domain.MulKkamResult
+import com.mulkkam.domain.model.MulKkamResult
 import com.mulkkam.domain.repository.NicknameRepository
 
 class NicknameRepositoryImpl(
