+++ android/app/src/main/java/com/mulkkam/data/repository/NotificationRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.mulkkam.data.remote.model.error.toResponseError
 import com.mulkkam.data.remote.model.request.ActiveCaloriesBurnedRequest
 import com.mulkkam.data.remote.model.response.toDomain
 import com.mulkkam.data.remote.service.NotificationsService
-import com.mulkkam.domain.MulKkamResult
+import com.mulkkam.domain.model.MulKkamResult
 import com.mulkkam.domain.model.Notification
 import com.mulkkam.domain.repository.NotificationRepository
 import java.time.LocalDateTime
