+++ android/app/src/main/java/com/mulkkam/domain/model/Gender.kt
@@ -1,4 +1,4 @@
-package com.mulkkam.domain
+package com.mulkkam.domain.model
 
 enum class Gender {
     MALE,
