+++ android/app/src/main/java/com/mulkkam/domain/model/MemberInfo.kt
@@ -1,6 +1,6 @@
 package com.mulkkam.domain.model
 
-import com.mulkkam.domain.Gender
+import com.mulkkam.domain.model.Gender
 
 data class MemberInfo(
     val nickname: String,
