+++ android/app/src/main/java/com/mulkkam/domain/model/Notification.kt
@@ -1,12 +1,12 @@
 package com.mulkkam.domain.model
 
-import com.mulkkam.domain.Alarm
+import com.mulkkam.domain.model.NotificationType
 import java.time.LocalDateTime
 
 data class Notification(
     val id: Int,
     val title: String,
-    val type: Alarm,
+    val type: NotificationType,
     val createdAt: LocalDateTime,
     val recommendedTargetAmount: Int,
     val isRead: Boolean,
