+++ android/app/src/main/java/com/mulkkam/domain/model/NotificationType.kt
@@ -1,14 +1,14 @@
-package com.mulkkam.domain
+package com.mulkkam.domain.model
 
-enum class Alarm {
+enum class NotificationType {
     SUGGESTION,
     REMIND,
     NOTICE,
     ;
 
     companion object {
-        fun from(alarm: String): Alarm =
-            when (alarm) {
+        fun from(notificationType: String): NotificationType =
+            when (notificationType) {
                 SUGGESTION.name -> SUGGESTION
                 REMIND.name -> REMIND
                 NOTICE.name -> NOTICE
