+++ android/app/src/main/java/com/mulkkam/domain/model/OnboardingInfo.kt
@@ -1,6 +1,6 @@
 package com.mulkkam.domain.model
 
-import com.mulkkam.domain.Gender
+import com.mulkkam.domain.model.Gender
 
 data class OnboardingInfo(
     val nickname: String? = null,
