+++ android/app/src/main/java/com/mulkkam/domain/model/TargetAmount.kt
@@ -1,7 +1,5 @@
 package com.mulkkam.domain.model
 
-import com.mulkkam.domain.MulKkamError
-
 @JvmInline
 value class TargetAmount(
     val amount: Int,
