+++ android/app/src/main/java/com/mulkkam/domain/repository/AuthRepository.kt
@@ -1,6 +1,6 @@
 package com.mulkkam.domain.repository
 
-import com.mulkkam.domain.MulKkamResult
+import com.mulkkam.domain.model.MulKkamResult
 
 interface AuthRepository {
     suspend fun postAuthKakao(oauthAccessToken: String): MulKkamResult<String>
