+++ android/app/src/main/java/com/mulkkam/domain/repository/NicknameRepository.kt
@@ -1,6 +1,6 @@
 package com.mulkkam.domain.repository
 
-import com.mulkkam.domain.MulKkamResult
+import com.mulkkam.domain.model.MulKkamResult
 
 interface NicknameRepository {
     suspend fun getNicknameValidation(nickname: String): MulKkamResult<Unit>
