+++ android/app/src/main/java/com/mulkkam/ui/history/HistoryViewModel.kt
@@ -6,9 +6,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
 import androidx.lifecycle.viewModelScope
 import com.mulkkam.di.RepositoryInjection
-import com.mulkkam.domain.IntakeHistory
-import com.mulkkam.domain.IntakeHistorySummaries
-import com.mulkkam.domain.IntakeHistorySummary
+import com.mulkkam.domain.model.IntakeHistory
+import com.mulkkam.domain.model.IntakeHistorySummaries
+import com.mulkkam.domain.model.IntakeHistorySummary
 import kotlinx.coroutines.launch
 import java.time.DayOfWeek
 import java.time.LocalDate
