+++ android/app/src/main/java/com/mulkkam/ui/onboarding/bioinfo/BioInfoViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.mulkkam.domain.Gender
+import com.mulkkam.domain.model.Gender
 
 class BioInfoViewModel : ViewModel() {
     private val _gender = MutableLiveData<Gender>()
