+++ android/app/src/main/java/com/mulkkam/ui/onboarding/targetamount/TargetAmountViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.mulkkam.di.RepositoryInjection
-import com.mulkkam.domain.Gender
+import com.mulkkam.domain.model.Gender
 import com.mulkkam.domain.model.TargetAmount
 import kotlinx.coroutines.launch
